kafka-producer/requirements.txt
kafka-python==2.0.2
google-cloud-bigquery==3.16.0
requests==2.31.0
python-dotenv==1.0.1
